now box moves up and down, spawn at the screen width, spawn one by one every specific time(differ by currentstage) / adjusted box and bullet's collider size / fixed some casting warnings / now we can get scoreboard's height(this would be useful to know the boundary)

+++ SDL2_project/Formation.cpp
@@ -43,7 +43,7 @@ void Formation::Update()
 		if (mOffsetTimer >= mOffsetDelay)
 		{
 			mOffsetCounter++;
-			Translate(VECTOR2D_UP * mOffsetDirection * mOffsetAmount, world);
+			Translate(VECTOR2D_UP * (float)mOffsetDirection * mOffsetAmount, world);
 
 			if (mOffsetCounter == 8)
 			{
